feat: add native macOS directory compare
